requirements.txt
langchain
langchain-community
langchain-huggingface
langgraph
faiss-cpu
pypdf
python-dotenv
sentence-transformers
langchain-google-genai
